Fix warnings on Visual Studio (#1431)
* Fix warnings C4244, C4267 in src/layer/yolov3detectionoutput.cpp

C4244: '=': conversion from 'int' to 'float', possible loss of data
C4244: 'initializing': conversion from 'float' to 'int', possible loss of data
C4244: 'initializing': conversion from 'double' to 'float', possible loss of data
C4244: 'return': conversion from 'double' to 'float', possible loss of data
C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warnings C4244, C4267 in src/layer/yolodetectionoutput.cpp

C4244: '=': conversion from 'int' to 'float', possible loss of data
C4244: 'initializing': conversion from 'float' to 'int', possible loss of data
C4244: 'initializing': conversion from 'double' to 'float', possible loss of data
C4244: 'return': conversion from 'double' to 'float', possible loss of data
C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4244 in src/layer/quantize.cpp

C4244: 'initializing': conversion from 'double' to 'int', possible loss of data

* Fix warnings C4244, C4267 in src/layer/detectionoutput.cpp

C4244: '=': conversion from 'int' to 'float', possible loss of data
C4244: 'initializing': conversion from 'double' to 'float', possible loss of data
C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4244 in src/layer/roipooling.cpp

C4244: 'initializing': conversion from 'double' to 'int', possible loss of data

* Fix warning C4244 in src/layer/sigmoid.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4267 in src/layer/slice.cpp

C4267: '=': conversion from 'size_t' to 'int', possible loss of data
C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4267 in src/layer/softmax.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/interp.cpp

C4244: '=': conversion from 'float' to 'int', possible loss of data
C4244: 'initializing': conversion from 'double' to 'int', possible loss of data

* Fix warning C4244 in src/layer/instancenorm.cpp

C4244: 'initializing': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/deconvolutiondepthwise.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/convolutiondepthwise.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/net.cpp

C4244: 'return': conversion from '__int64' to 'int', possible loss of data
C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
C4267: 'return': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4244 in src/layer/bnll.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4267 in src/layer/concat.cpp

C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4267 in tools/mxnet/mxnet2ncnn.cpp

C4244: 'initializing': conversion from 'double' to 'float', possible loss of data
C4267: '=': conversion from 'size_t' to 'int', possible loss of data
C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
C4305: 'initializing': truncation from 'double' to 'float'

+++ src/layer/deconvolutiondepthwise.cpp
@@ -197,7 +197,7 @@ int DeconvolutionDepthWise::forward(const Mat& bottom_blob, Mat& top_blob, const
 
                 for (int i = 0; i < size; i++)
                 {
-                    outptr[i] = 1.f / (1.f + exp(-outptr[i]));
+                    outptr[i] = static_cast<float>(1.f / (1.f + exp(-outptr[i])));
                 }
             }
         }
@@ -291,7 +291,7 @@ int DeconvolutionDepthWise::forward(const Mat& bottom_blob, Mat& top_blob, const
 
                     for (int i = 0; i < size; i++)
                     {
-                        outptr[i] = 1.f / (1.f + exp(-outptr[i]));
+                        outptr[i] = static_cast<float>(1.f / (1.f + exp(-outptr[i])));
                     }
                 }
             }
